Make both branches of if in socket_tcp.hpp the same type.

+++ shared/net/transport_layer/socket_tcp.hpp
@@ -11,7 +11,7 @@
 
 #define KP(fmt, ...) \
     do { kprintf(fmt, ##__VA_ARGS__); } while (0)
-    
+
 extern "C" {
     void      sleep(uint64_t ms);
     uintptr_t malloc(uint64_t size);
@@ -154,7 +154,7 @@ public:
         flow->payload = { (uintptr_t)buf, (uint32_t)len };
         flow->flags = (1<<PSH_F) | (1<<ACK_F);
         tcp_result_t res = tcp_flow_send(flow);
-        return (res == TCP_OK) ? (int64_t)len : res;
+        return (res == TCP_OK) ? (int64_t)len : (int64_t)res;
     }
 
     int64_t recv(void* buf, uint64_t len) {
@@ -188,7 +188,7 @@ public:
     }
 
     int32_t close_client() {
-        
+
         if (connected && flow) {
             tcp_flow_close(flow);
             connected = false;
